Pre Release, Finance Feature

+++ app/src/main/java/com/example/csks_creatives/presentation/taskDetailScreen/components/AmountSection.kt
@@ -87,7 +87,7 @@ fun AmountSection(
 
                     OutlinedTextField(
                         value = taskState.taskPartialPaymentsAmount.toString(),
-                        onValueChange = { onEvent(TaskDetailEvent.TaskPartialPaymentAmountChanged(it.toInt())) },
+                        onValueChange = { onEvent(TaskDetailEvent.TaskPartialPaymentAmountChanged(it)) },
                         label = { Text("Enter Partial Amount") },
                         singleLine = true,
                         modifier = Modifier.fillMaxWidth(),
